+++ tests/testthat/test-ot_indices_1d.R
@@ -4,7 +4,7 @@ test_that("ot_indices_1d handles univariate output", {
 
   expect_s3_class(result, "gsaot_indices")
   expect_named(result, c("method", "indices", "bound", "x", "y",
-                         "separation_measures", "partitions", "boot"))
+                         "separation_measures", "partitions", "is_L22", "boot"))
   expect_true(all(paste0("X", 1:8) %in% rownames(result$indices)))
 })
 
@@ -14,7 +14,7 @@ test_that("ot_indices_1d handles different p", {
 
   expect_s3_class(result, "gsaot_indices")
   expect_named(result, c("method", "indices", "bound", "x", "y",
-                         "separation_measures", "partitions", "boot"))
+                         "separation_measures", "partitions", "is_L22", "boot"))
   expect_true(all(paste0("X", 1:3) %in% names(result$indices)))
 })
 
@@ -25,7 +25,7 @@ test_that("ot_indices_1d handles bootstrapping", {
 
   expect_s3_class(result, "gsaot_indices")
   expect_named(result, c("method", "indices", "bound", "x", "y",
-                         "separation_measures", "partitions", "boot",
+                         "separation_measures", "partitions", "is_L22", "boot",
                          "indices_ci", "separation_measures_ci",
                          "bound_ci", "R", "type", "conf", "W_boot"))
   expect_true(all(paste0("X", 1:8) %in% rownames(result$indices)))